Attempt to use nodeless shaders. Then attempt to create a custom OSL shader.
Attempted to use non-node material but there was no support for phong properties such as specular colour and emissive colour. Began trying to put together an OSL shader as a candidate. Ran into problems with the interaction between closures (required for diffuse lighting) and the lack of support for computing specularity without a specular ramp. It was messy and rendering would only work with the CPU.